moved some api classes into sub packages
