+++ src/main/java/net/pms/network/mediaserver/handlers/api/starrating/RequestVO.java
@@ -1,4 +1,4 @@
-package net.pms.network.mediaserver.handlers.api;
+package net.pms.network.mediaserver.handlers.api.starrating;
 
 public class RequestVO {
